Move downsample_array to array_utils
Move associated tests to test/test_array_utils
