+++ tests/transforms/test_array_utils.py
@@ -234,3 +234,99 @@ def test_compare_crop_pad_and_extents(arr, shape):
             arr[extents[0]:extents[1], extents[2]:extents[3]],
             pad_width)
     assert np.all(cropped_padded == indexed)
+
+
+@pytest.mark.parametrize(
+        ("array, input_fps, output_fps, random_seed, strategy, expected"),
+        [
+            (
+                # random downsample 1D array
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 2, 0, 'random',
+                np.array([2, 5])),
+            (
+                # random downsample ND array
+                np.array([
+                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
+                    [5, 8], [11, 12]]),
+                7, 2, 0, 'random',
+                np.array([[2, 1], [5, 8]])),
+            (
+                # first downsample 1D array
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 2, 0, 'first',
+                np.array([1, 3])),
+            (
+                # random downsample ND array
+                np.array([
+                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
+                    [5, 8], [11, 12]]),
+                7, 2, 0, 'first',
+                np.array([[1, 3], [3, 2]])),
+            (
+                # last downsample 1D array
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 2, 0, 'last',
+                np.array([2, 11])),
+            (
+                # last downsample ND array
+                np.array([
+                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
+                    [5, 8], [11, 12]]),
+                7, 2, 0, 'last',
+                np.array([[2, 1], [11, 12]])),
+            (
+                # average downsample 1D array
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 2, 0, 'average',
+                np.array([13/4, 19/3])),
+            (
+                # average downsample ND array
+                np.array([
+                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
+                    [5, 8], [11, 12]]),
+                7, 2, 0, 'average',
+                np.array([[13/4, 4], [19/3, 22/3]])),
+            (
+                # maximum downsample 1D array
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 2, 0, 'maximum',
+                np.array([6, 11])),
+            ])
+def test_downsample(array, input_fps, output_fps, random_seed, strategy,
+                    expected):
+    array_out = au.downsample_array(
+            array=array,
+            input_fps=input_fps,
+            output_fps=output_fps,
+            strategy=strategy,
+            random_seed=random_seed)
+    assert np.array_equal(expected, array_out)
+
+
+@pytest.mark.parametrize(
+        ("array, input_fps, output_fps, random_seed, strategy, expected"),
+        [
+            (
+                # upsampling not defined
+                np.array([1, 4, 6, 2, 3, 5, 11]),
+                7, 11, 0, 'maximum',
+                np.array([6, 11])),
+            (
+                # maximum downsample ND array
+                # not defined
+                np.array([
+                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
+                    [5, 1234], [11, 12]]),
+                7, 2, 0, 'maximum',
+                np.array([[6, 8], [11, 12]])),
+            ])
+def test_downsample_exceptions(array, input_fps, output_fps, random_seed,
+                               strategy, expected):
+    with pytest.raises(ValueError):
+        au.downsample_array(
+                array=array,
+                input_fps=input_fps,
+                output_fps=output_fps,
+                strategy=strategy,
+                random_seed=random_seed)
