+++ tests/transforms/test_video_utils.py
@@ -22,74 +22,6 @@ def video_fixture():
 
 
 @pytest.mark.parametrize(
-        ("array, input_fps, output_fps, random_seed, strategy, expected"),
-        [
-            (
-                # random downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'random',
-                np.array([2, 5])),
-            (
-                # random downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'random',
-                np.array([[2, 1], [5, 8]])),
-            (
-                # first downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'first',
-                np.array([1, 3])),
-            (
-                # random downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'first',
-                np.array([[1, 3], [3, 2]])),
-            (
-                # last downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'last',
-                np.array([2, 11])),
-            (
-                # last downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'last',
-                np.array([[2, 1], [11, 12]])),
-            (
-                # average downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'average',
-                np.array([13/4, 19/3])),
-            (
-                # average downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'average',
-                np.array([[13/4, 4], [19/3, 22/3]])),
-            (
-                # maximum downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'maximum',
-                np.array([6, 11])),
-            ])
-def test_downsample(array, input_fps, output_fps, random_seed, strategy,
-                    expected):
-    array_out = transformations.downsample_array(
-            array=array,
-            input_fps=input_fps,
-            output_fps=output_fps,
-            strategy=strategy,
-            random_seed=random_seed)
-    assert np.array_equal(expected, array_out)
-
-
-@pytest.mark.parametrize(
         ("array, input_fps, output_fps, strategy, expected"),
         [
             (
@@ -121,34 +53,6 @@ def test_video_downsample(
             strategy)
 
     assert np.array_equal(downsampled_video, expected)
-
-
-@pytest.mark.parametrize(
-        ("array, input_fps, output_fps, random_seed, strategy, expected"),
-        [
-            (
-                # upsampling not defined
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 11, 0, 'maximum',
-                np.array([6, 11])),
-            (
-                # maximum downsample ND array
-                # not defined
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 1234], [11, 12]]),
-                7, 2, 0, 'maximum',
-                np.array([[6, 8], [11, 12]])),
-            ])
-def test_downsample_exceptions(array, input_fps, output_fps, random_seed,
-                               strategy, expected):
-    with pytest.raises(ValueError):
-        transformations.downsample_array(
-                array=array,
-                input_fps=input_fps,
-                output_fps=output_fps,
-                strategy=strategy,
-                random_seed=random_seed)
 
 
 def test_mp4_conversion(video_fixture, tmp_path):
